pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = ""
authors = [
    {name = "kawa",email = "moshimoshi@yahoo.de"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=12.1.1,<13.0.0)"
]

[tool.poetry]
packages = [{include = "ocr", from = "src"},
			{include = "data", from = "src"}
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
